requirements.txt
about-time==4.2.1
alive-progress==3.0.1
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
distlib==0.3.6
elastic-transport==8.4.0
elasticsearch==8.6.0
filelock==3.9.0
fonttools==4.38.0
grapheme==0.6.0
huggingface-hub==0.11.1
identify==2.5.18
idna==3.4
img2vec-pytorch==1.0.1
importlib-resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.0
nltk==3.8.1
nodeenv==1.7.0
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==22.0
Pillow==9.4.0
platformdirs==3.0.0
pre-commit==3.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
scikit-learn==1.2.0
scipy==1.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
urllib3==1.26.13
virtualenv==20.20.0
zipp==3.15.0
